Only thread the unlinking part of the jobs instead of the job_purge. This avoids segfaulting. Also, use a threadpool instead of creating a thread each time in order to increase speed

+++ src/resmom/job_func.c
@@ -176,6 +176,7 @@ extern tlist_head svr_alljobs;
 
 void nodes_free(job *);
 int TTmpDirName(job *, char *);
+void *purge_file(void *);
 
 extern int multi_mom;
 extern int pbs_rm_port;
@@ -247,9 +248,12 @@ int remtree(
   DIR  *dir;
 
   struct dirent *pdir;
-  char  namebuf[MAXPATHLEN], *filnam;
-  int  i;
-  int  rtnv = 0;
+  char           namebuf[MAXPATHLEN];
+  char          *filnam;
+  char          *alloced_path;
+  char          *alloced_dir;
+  int            i;
+  int            rtnv = 0;
 #if defined(HAVE_STRUCT_STAT64) && defined(HAVE_STAT64) && defined(LARGEFILE_WORKS)
 
   struct stat64 sb;
@@ -271,6 +275,8 @@ int remtree(
 
     return(-1);
     }
+
+  alloced_dir = strdup(dirname);
 
   if (S_ISDIR(sb.st_mode))
     {
@@ -311,23 +317,15 @@ int remtree(
         continue;
         }
 
+      alloced_path = strdup(namebuf);
+
       if (S_ISDIR(sb.st_mode))
         {
         rtnv = remtree(namebuf);
         }
-      else if (unlink(namebuf) < 0)
-        {
-        if (errno != ENOENT)
-          {
-          sprintf(log_buffer, "unlink failed on %s",
-                  namebuf);
-
-          log_err(errno, id, log_buffer);
-
-          rtnv = -1;
-          }
-        }
-      else if (LOGLEVEL >= 7)
+      else if ((alloced_path != NULL) &&
+               ((enqueue_threadpool_request(purge_file,alloced_path)) == PBSE_NONE) &&
+               (LOGLEVEL >= 7))
         {
         sprintf(log_buffer, "unlink(1) succeeded on %s", namebuf);
 
@@ -356,16 +354,9 @@ int remtree(
       log_ext(-1, id, log_buffer, LOG_DEBUG);
       }
     }
-  else if (unlink(dirname) < 0)
-    {
-    sprintf(log_buffer, "unlink failed on %s",
-            dirname);
-
-    log_err(errno, id, log_buffer);
-
-    rtnv = -1;
-    }
-  else if (LOGLEVEL >= 7)
+  else if ((alloced_dir != NULL) &&
+           (enqueue_threadpool_request(purge_file,alloced_dir) == PBSE_NONE) &&
+           (LOGLEVEL >= 7))
     {
     sprintf(log_buffer, "unlink(2) succeeded on %s", dirname);
 
@@ -621,28 +612,87 @@ static void job_init_wattr(
 
 
 
+void *purge_file(
 
-/*
- * job_purge_thread - purge job from system
- *
- * The job is dequeued; the job control file, script file and any spooled
- * output files are unlinked, and the job structure is freed.
- * If we are MOM, the task files and checkpoint files are also
- * removed.
- */
+  void *vp)
 
-void *job_purge_thread(void *jobtopurge)
   {
-  static char   id[] = "job_purge_thread";
-  job *pjob;
+  static char *id = "purge_file";
+  char  *path = (char *)vp;
+
+  if (unlink(path) < 0)
+    {
+    if (errno != ENOENT)
+      {
+      snprintf(log_buffer,sizeof(log_buffer),
+        "Counldn't remove %s",
+        path);
+      log_err(errno, id, log_buffer);
+      }
+    }
+  else if (LOGLEVEL >= 6)
+    {
+    snprintf(log_buffer, sizeof(log_buffer),
+      "Successfully removed %s",
+      path);
+
+    log_record(PBSEVENT_DEBUG,PBS_EVENTCLASS_JOB,"",log_buffer);
+    }
+
+  free(path);
+
+  return(NULL);
+  } /* END purge_file() */
+
+
+
+
+void *remove_dir(
+
+  void *vp)
+
+  {
+  static char *id = "remove_dir";
+  char        *path = (char *)vp;
+
+  if (rmdir(path) != 0)
+    {
+    if (errno != ENOENT)
+      {
+      snprintf(log_buffer,sizeof(log_buffer),
+        "Couldn't remove %s",
+        path);
+      log_err(errno,id,log_buffer);
+      }
+    else if (LOGLEVEL >= 6)
+      {
+      snprintf(log_buffer,sizeof(log_buffer),
+        "Successfully removed %s",
+        path);
+
+      log_record(PBSEVENT_JOB,PBS_EVENTCLASS_JOB,"",log_buffer);
+      }
+    }
+  } /* END remove_dir() */
+
+
+
+
+
+
+void job_purge(
+
+  job *pjob)  /* I (modified) */
+
+  {
+  static char   id[] = "job_purge";
   char          namebuf[MAXPATHLEN + 1];
+  char         *alloced_path;
   char          portname[MAXPATHLEN + 1];
   extern char  *msg_err_purgejob;
   int           rc;
 
   extern void MOMCheckRestart(void);
-
-  pjob = (job *)jobtopurge;
 
   if (pjob->ji_flags & MOM_HAS_TMPDIR)
     {
@@ -661,8 +711,7 @@ void *job_purge_thread(void *jobtopurge)
           (seteuid(pjob->ji_qs.ji_un.ji_momt.ji_exuid) == -1))
         {
         /* FAILURE */
-
-        pthread_exit(NULL);
+        return;
         }
 
       rc = remtree(namebuf);
@@ -670,7 +719,8 @@ void *job_purge_thread(void *jobtopurge)
       seteuid(pbsuser);
       setegid(pbsgroup);
 
-      if ((rc != 0) && (LOGLEVEL >= 5))
+      if ((rc != 0) && 
+          (LOGLEVEL >= 5))
         {
         sprintf(log_buffer,
           "recursive remove of job transient tmpdir %s failed",
@@ -691,7 +741,6 @@ void *job_purge_thread(void *jobtopurge)
     extern void delete_cpuset(char *);
 
     /* Delete the cpuset for the job. */
-
     delete_cpuset(pjob->ji_qs.ji_jobid);
     }
 
@@ -707,7 +756,15 @@ void *job_purge_thread(void *jobtopurge)
       path_aux,
       pjob->ji_qs.ji_jobid);
 
-    unlink(file);
+    alloced_path = strdup(file);
+    if (file != NULL)
+      {
+      enqueue_threadpool_request(purge_file,alloced_path);
+      }
+    else
+      {
+      log_err(ENOMEM,id,"Cannot allocate memory to purge file");
+      }
 
     pjob->ji_flags &= ~MOM_HAS_NODEFILE;
 
@@ -717,7 +774,15 @@ void *job_purge_thread(void *jobtopurge)
       path_aux,
       pjob->ji_qs.ji_jobid);
 
-    unlink(file);
+    alloced_path = strdup(file);
+    if (file != NULL)
+      {
+      enqueue_threadpool_request(purge_file,alloced_path);
+      }
+    else
+      {
+      log_err(ENOMEM,id,"Cannot allocate memory to purge file");
+      }
     }
 
   delete_link(&pjob->ji_jobque);
@@ -747,22 +812,15 @@ void *job_purge_thread(void *jobtopurge)
 
   strcat(namebuf,JOB_SCRIPT_SUFFIX);
 
-  if (unlink(namebuf) < 0)
+  alloced_path = strdup(namebuf);
+  if (alloced_path == NULL)
     {
-    if (errno != ENOENT)
-      log_err(errno, id, msg_err_purgejob);
+    log_err(ENOMEM,id,"Cannot allocate space to purge file");
     }
-  else if (LOGLEVEL >= 6)
+  else
     {
-    sprintf(log_buffer, "removed job script");
-
-    log_record(
-      PBSEVENT_DEBUG,
-      PBS_EVENTCLASS_JOB,
-      pjob->ji_qs.ji_jobid,
-      log_buffer);
+    enqueue_threadpool_request(purge_file,alloced_path);
     }
-
 
 #if IBM_SP2==2        /* IBM SP PSSP 3.1 */
   unload_sp_switch(pjob);
@@ -796,20 +854,14 @@ void *job_purge_thread(void *jobtopurge)
 
   strcat(namebuf, JOB_FILE_SUFFIX);
 
-  if (unlink(namebuf) < 0)
+  alloced_path = strdup(namebuf);
+  if (alloced_path == NULL)
     {
-    if (errno != ENOENT)
-      log_err(errno, id, msg_err_purgejob);
+    log_err(ENOMEM,id,"Cannot allocate space to purge file");
     }
-  else if (LOGLEVEL >= 6)
+  else
     {
-    sprintf(log_buffer, "removed job file");
-
-    log_record(
-      PBSEVENT_DEBUG,
-      PBS_EVENTCLASS_JOB,
-      pjob->ji_qs.ji_jobid,
-      log_buffer);
+    enqueue_threadpool_request(purge_file,alloced_path);
     }
 
   job_free(pjob);
@@ -818,50 +870,6 @@ void *job_purge_thread(void *jobtopurge)
 
   if (((job *)GET_NEXT(svr_alljobs)) == NULL)
     MOMCheckRestart();
-
-  pthread_exit(NULL);
-  }
-
-void job_purge(
-
-  job *pjob)  /* I (modified) */
-
-  {
-  static char   id[] = "job_purge";
-  pthread_attr_t attr;
-  pthread_t *thread;
-  int rc;
-
-  rc = pthread_attr_init( &attr );
-  if(rc)
-    {
-    sprintf(log_buffer, "pthread_attr_init failded. job number %s", pjob->ji_qs.ji_jobid);
-    log_err(rc, id, log_buffer);
-    return;
-    }
-
-  rc = pthread_attr_setdetachstate(
-     &attr, PTHREAD_CREATE_DETACHED );
-  if(rc)
-    {
-    sprintf(log_buffer, "pthread_attr_setdetachstate failded. job number %s", pjob->ji_qs.ji_jobid);
-    log_err(rc, id, log_buffer);
-    return;
-    }
-  thread = (pthread_t *)malloc(sizeof(pthread_t));
-  if(thread == NULL) 
-    {
-    sprintf(log_buffer, "Could not allocate memory for job_purge thread");
-    log_err(errno, id, log_buffer);
-    return;
-    }
-
-  rc = pthread_create(thread, &attr, &job_purge_thread, pjob);
-  if(rc)
-    {
-    sprintf(log_buffer, "pthread_create failed: %d", rc);
-    log_err(rc, id, log_buffer);
-    }
 
   return;
   }  /* END job_purge() */
